refactor: relocating components
